requirements.txt
fabric3
Jinja2